move logger utilities
