+++ src/pytorch_lightning/callbacks/progress/base.py
@@ -14,8 +14,8 @@
 from typing import Any, Dict, Optional, Union
 
 import pytorch_lightning as pl
+from lightning_fabric.utilities.logger import _version
 from pytorch_lightning.callbacks import Callback
-from pytorch_lightning.utilities.logger import _version
 from pytorch_lightning.utilities.rank_zero import rank_zero_warn
