+++ src/pytorch_lightning/loggers/comet.py
@@ -25,9 +25,9 @@ from lightning_utilities.core.imports import module_available
 from torch import Tensor
 
 import pytorch_lightning as pl
+from lightning_fabric.utilities.logger import _add_prefix, _convert_params, _flatten_dict
 from pytorch_lightning.loggers.logger import Logger, rank_zero_experiment
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.logger import _add_prefix, _convert_params, _flatten_dict
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
 
 log = logging.getLogger(__name__)
