+++ src/pytorch_lightning/loggers/tensorboard.py
@@ -27,12 +27,12 @@ from torch import Tensor
 
 import pytorch_lightning as pl
 from lightning_fabric.utilities.cloud_io import get_filesystem
+from lightning_fabric.utilities.logger import _add_prefix, _convert_params, _flatten_dict
+from lightning_fabric.utilities.logger import _sanitize_params as _utils_sanitize_params
 from lightning_fabric.utilities.types import _PATH
 from pytorch_lightning.core.saving import save_hparams_to_yaml
 from pytorch_lightning.loggers.logger import Logger, rank_zero_experiment
 from pytorch_lightning.utilities.imports import _OMEGACONF_AVAILABLE
-from pytorch_lightning.utilities.logger import _add_prefix, _convert_params, _flatten_dict
-from pytorch_lightning.utilities.logger import _sanitize_params as _utils_sanitize_params
 from pytorch_lightning.utilities.rank_zero import rank_zero_only, rank_zero_warn
 
 log = logging.getLogger(__name__)
